refactor(opencode): extract the prompt delivery attempt text builders
Problem: `OpenCodePromptDeliveryReadCommitPolicy.ts` is the read-commit
decision module, but it had also accumulated the builders that compose the
text a delivery attempt actually carries. Those two concerns answer
different questions - "may this response close the inbox row?" versus "what
does the next attempt say?" - and the file is the one module in this folder
with no frozen legacy cap, so every line it grows is spent against the plain
800-line limit that the upcoming turn-activity work needs.

Fix: move `getOpenCodeDeliveryHardFailureKind`,
`buildOpenCodePromptDeliveryRepairControlText` and
`buildOpenCodePromptDeliveryAttemptText` verbatim into a sibling
`OpenCodePromptDeliveryAttemptText.ts` and point the single caller, the
member message delivery service, at the new path. No behaviour changes; the
moved bodies are byte-identical.

Why this shape: the new module sits next to the file it came from, is named
after the concern, and exports only the entry points the delivery service
calls, matching the extraction pattern already used for
`OpenCodeLegacyMemberMessageDelivery.ts` and
`OpenCodePromptDeliveryStalePendingPolicy.ts` in the same folder. The
dependency runs one way - attempt text reads `hasOpenCodeNonVisibleProgressProof`
from the read-commit policy, never the reverse - so no re-export shim is
needed and the import graph stays acyclic.
